Improve error handling in text dump
Turns out the value of the error running `dpkg` is not being used as the value being shipped to coalmine, it was a visual artifact of bad error handling where errors were inserted inline carelessly. So now, we report errors prior to outputting the dump, and insert "unknown" in place of the error.