Fixed issue with .rpmsave files
when uprgading from an agent.conf to an agent.yml system, RPM based distros will
kindly save the conf file as agent.conf.rpmsave during the upgrade process. I
can't seem to find a way to fix this since the old RPM specifies that file as a
config file.

When converting files we now check for .rpmsave as well as the original .conf versions.